+++ src/algo/mod.rs
@@ -1,4 +1,6 @@
 //! Graph algorithms
+//!
+//! Traverse and search graphs.
 
 pub mod bfs;
 pub mod dijkstra;