+++ src/gen/mod.rs
@@ -4,8 +4,10 @@
 
 pub mod cycle;
 pub mod linear;
+pub mod linear_const;
 
 pub use {
     cycle::Cycle,
     linear::Linear,
+    linear_const::LinearConst,
 };